backend/requirements.txt
# Core dependencies
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9

# Image processing
Pillow==9.5.0
opencv-python==4.8.1.78
numpy==1.24.3

# Face detection
torch==2.1.2
torchvision==0.16.2
facenet-pytorch==2.5.3

# Utilities
python-jose==3.3.0
python-dotenv==1.0.0 

ExifRead==3.0.0
piexif==1.1.3